ref(processort): Move user and client reports processing into separate submodule
